requirements.txt
certifi==2025.6.15
charset-normalizer==3.4.2
contourpy==1.3.0
cycler==0.12.1
fonttools==4.58.4
idna==3.10
importlib_resources==6.5.2
kiwisolver==1.4.7
matplotlib==3.9.4
numpy==2.0.2
packaging==25.0
pillow==11.3.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
requests==2.32.4
six==1.17.0
urllib3==2.5.0
zipp==3.23.0
